Add files via upload
This project is a website for the Girl Child Advocacy (GCA) NGO, an organization dedicated to empowering and advocating for the rights and education of young girls, particularly those who are orphaned or have dropped out of school in Northern Ghana. The website provides information about the organization's mission, programs, team, and ways to get involved.